didispeech_gui, dialog, input_file: add select file support
Add select file support both for input file and output file.
Add class InputFile to handle input_file related stuffs.
When you push select file button, a dialog will open and
allows you to select the file. If you are choosing the input
file, only supported format are showed.
When you choose a file, it will be written as text of the button.
If no file is selected, the text stay still.