didi, multithread, gui, utils, inputfile: add the function to perform the transcription
At this point, the application run. Improvements are needed
but the basics work well.

Add the main class of the application: Didi.
Didi handles the main functions to perform the parsing.
It load the audio file, splits it into chunks, gives them
to several threads and starts them.

Multithread class also in very important. It is used to
start Didi in another thread so the gui remains responsive
meanwhile the transcription is running.

In utils there are some useful miscelanus functions.

Other minor improvement in the other files.