gui, multithread: add printing load
Add the printing of "load" when a long process is performing.
When a long process has begun, a timer starts (in multithread class);
this timer emits a signal every a while (currently 1 second).
At every signal, a "loading..." is written. Actually, at every signal
a "." is added to the text.
When the process end, the timer is stopped and "Done!" is written.