didispeech: add validation to chunk_size option
chunk size can be 1000 to 60000.
At least 1000 because performance reasons; max 60000 because google
allow to transcript up to 60sec.

User is bounded to these values by a regex which limit the _e_chunk_size
edit line of the advanced option;
Furthermore, in core/didi.py there is a check: if chunk size is out of the
boundaries, it is set to 1000 or 60000.